WIP: Add support for grouped phases parsing
Phases are now parsed as a group and are a sub-tree. E.g.
['CHANGE_STATUS', 'PHASE', ['LIQ', 'FCC'], 'DORMANT']
This is applied everywhere

Also adds an extension by considering a property to be a thermodynamic value with
a list of phases e.g.
X(LIQ,FCC) or MUR(LIQ,FCC)
This is parsed as
['X', ['LIQ', 'FCC']]

Other minor improvements with a little refactoring, including not adding the
extra symbols to the 'symbol_name' to ref_phase_name to separate these.